up: update some logic for storage,restore ws connection
