+++ src/websocket-server/src/ConnectionStorage.php
@@ -4,8 +4,6 @@ namespace Swoft\WebSocket\Server;
 
 use Swoft;
 use Swoft\Bean\Annotation\Mapping\Bean;
-use Swoft\Http\Message\Request as Psr7Request;
-use Swoft\Http\Message\Response as Psr7Response;
 use Swoft\Session\Session;
 use Swoft\Stdlib\Helper\JsonHelper;
 use Swoft\WebSocket\Server\Contract\StorageInterface;
@@ -38,14 +36,15 @@ class ConnectionStorage
 
     /**
      * Storage connection on handshake successful
+     *
+     * @param Request  $request
+     *
      * @see HandshakeListener::onHandshake()
      *
      * You should call the method on event: {@see WsServerEvent::HANDSHAKE_SUCCESS}
      *
-     * @param Request  $request
-     * @param Response $response
      */
-    public function storage(Request $request, Response $response): void
+    public function storage(Request $request): void
     {
         if (!$this->enable) {
             return;
@@ -54,31 +53,19 @@ class ConnectionStorage
         $key  = self::genKey($fd = $request->fd);
         $conn = Session::mustGet($fd);
 
-        $this->storage->set($key, JsonHelper::encode([
-            // request data
-            'fd'        => $fd,
-            'get'       => $request->get,
-            'post'      => $request->post,
-            'cookie'    => $request->cookie,
-            'header'    => $request->header,
-            'server'    => $request->server,
-            // response data
-            'resHeader' => $response->header,
-            'resCookie' => $response->cookie,
-            // module info
-            'moduleInfo' => $conn->getModuleInfo()
-        ]));
+        $this->storage->set($key, $conn->toString());
     }
 
     /**
      * Restore connection on worker reload
-     * @see MessageListener::onMessage()
-     *
-     * You should call the method on event: {@see WsServerEvent::MESSAGE_RECEIVE}
      *
      * @param int $fd
      *
      * @return bool
+     * @see MessageListener::onMessage()
+     *
+     * You should call the method on event: {@see WsServerEvent::MESSAGE_RECEIVE}
+     *
      */
     public function restore(int $fd): bool
     {
@@ -88,38 +75,14 @@ class ConnectionStorage
 
         $key = self::genKey($fd);
 
-        // if not exist
+        // If not exist
         if (!$json = $this->storage->get($key)) {
             return false;
         }
 
-        $data = JsonHelper::decode($json);
-
-        // New request and response
-        $req = new Request();
-        $res = new Response();
-
-        // Init request
-        $req->fd     = $fd;
-        $req->get    = $data['get'];
-        $req->post   = $data['post'];
-        $req->cookie = $data['cookie'];
-        $req->header = $data['header'];
-        $req->server = $data['server'];
-
-        // Init response
-        $res->cookie = $data['resCookie'];
-        $res->header = $data['resHeader'];
-
-        // Initialize psr7 Request and Response
-        $psr7Req  = Psr7Request::new($req);
-        $psr7Res  = Psr7Response::new($res);
-        $wsServer = Swoft::getBean('wsServer');
-
         // Restore connection object
-        $conn = Connection::new($wsServer, $psr7Req, $psr7Res);
-        $conn->setHandshake(true);
-        $conn->setModuleInfo($data['moduleInfo']);
+        $data = JsonHelper::decode($json);
+        $conn = Connection::newFromArray($data);
 
         // Bind connection and bind cid => sid(fd)
         Session::set((string)$fd, $conn);
@@ -129,13 +92,14 @@ class ConnectionStorage
 
     /**
      * Remove storage connection data on close connection
-     * @see CloseListener::onClose()
-     *
-     * You should call the method on event: {@see WsServerEvent::CLOSE_BEFORE}
      *
      * @param int $fd
      *
      * @return bool
+     * @see CloseListener::onClose()
+     *
+     * You should call the method on event: {@see WsServerEvent::CLOSE_BEFORE}
+     *
      */
     public function remove(int $fd): bool
     {
